requirements.txt
pandas>=1.5
pyyaml>=5.4
pytest>=6.0